ingest/requirements.txt
psycopg2-binary
pandas
python-dotenv
requests
black
pylint
aiohttp
sqlalchemy[asyncio]